fix(runs): satisfy audit on the new corpus primitives
- Move runs query/drift integration tests into runs/corpus_tests.rs so
  runs.rs stays under the 1500-line god_file threshold (1322 lines now).
- Replace the public scalar_to_string helper with a private scalar_label
  closure inside common.rs. Same behavior, no unreferenced public export
  and no parallel-implementation flag against value_type_name in
  src/core/config.rs.

Audit reports 0 findings on the changed-since=main scope and PASSED.
